+++ spec/guard/dsl_spec.rb
@@ -107,7 +107,7 @@ describe Guard::Dsl do
     it 'adds multiple notification to the notifier' do
       ::Guard::Notifier.should_receive(:add_notifier).with(:growl, { silent: false })
       ::Guard::Notifier.should_receive(:add_notifier).with(:ruby_gntp, { host: '192.168.1.5', silent: false })
-      described_class.evaluate_guardfile(guardfile_contents: "notification :growl\nnotification :ruby_gntp, :host => '192.168.1.5'")
+      described_class.evaluate_guardfile(guardfile_contents: "notification :growl\nnotification :ruby_gntp, host: '192.168.1.5'")
     end
   end
 
@@ -127,7 +127,7 @@ describe Guard::Dsl do
 
     it 'passes the options to the interactor' do
       ::Guard::UI.should_not_receive(:deprecation).with(Guard::Deprecator::DSL_METHOD_INTERACTOR_DEPRECATION)
-      described_class.evaluate_guardfile(guardfile_contents: 'interactor :option1 => \'a\', :option2 => 123')
+      described_class.evaluate_guardfile(guardfile_contents: 'interactor option1: \'a\', option2: 123')
       Guard::Interactor.options.should include({ option1: 'a', option2: 123 })
     end
   end
@@ -184,7 +184,7 @@ describe Guard::Dsl do
     it 'receives options when specified, from normal arg' do
       ::Guard.should_receive(:add_plugin).with('test', { watchers: [], callbacks: [], opt_a: 1, opt_b: 'fancy', group: :default })
 
-      described_class.evaluate_guardfile(guardfile_contents: 'guard \'test\', :opt_a => 1, :opt_b => \'fancy\'')
+      described_class.evaluate_guardfile(guardfile_contents: 'guard \'test\', opt_a: 1, opt_b: \'fancy\'')
     end
   end
 
@@ -236,52 +236,52 @@ describe Guard::Dsl do
 
     context 'with valid options' do
       it 'sets the logger log level' do
-        described_class.evaluate_guardfile(guardfile_contents: 'logger :level => :error')
+        described_class.evaluate_guardfile(guardfile_contents: 'logger level: :error')
         Guard::UI.options.level.should eq :error
       end
 
       it 'sets the logger log level and convert to a symbol' do
-        described_class.evaluate_guardfile(guardfile_contents: 'logger :level => \'error\'')
+        described_class.evaluate_guardfile(guardfile_contents: 'logger level: \'error\'')
         Guard::UI.options.level.should eq :error
       end
 
       it 'sets the logger template' do
-        described_class.evaluate_guardfile(guardfile_contents: 'logger :template => \':message - :severity\'')
+        described_class.evaluate_guardfile(guardfile_contents: 'logger template: \':message - :severity\'')
         Guard::UI.options.template.should eq ':message - :severity'
       end
 
       it 'sets the logger time format' do
-        described_class.evaluate_guardfile(guardfile_contents: 'logger :time_format => \'%Y\'')
+        described_class.evaluate_guardfile(guardfile_contents: 'logger time_format: \'%Y\'')
         Guard::UI.options.time_format.should eq '%Y'
       end
 
       it 'sets the logger only filter from a symbol' do
-        described_class.evaluate_guardfile(guardfile_contents: 'logger :only => :cucumber')
+        described_class.evaluate_guardfile(guardfile_contents: 'logger only: :cucumber')
         Guard::UI.options.only.should eq(/cucumber/i)
       end
 
       it 'sets the logger only filter from a string' do
-        described_class.evaluate_guardfile(guardfile_contents: 'logger :only => \'jasmine\'')
+        described_class.evaluate_guardfile(guardfile_contents: 'logger only: \'jasmine\'')
         Guard::UI.options.only.should eq(/jasmine/i)
       end
 
       it 'sets the logger only filter from an array of symbols and string' do
-        described_class.evaluate_guardfile(guardfile_contents: 'logger :only => [:rspec, \'cucumber\']')
+        described_class.evaluate_guardfile(guardfile_contents: 'logger only: [:rspec, \'cucumber\']')
         Guard::UI.options.only.should eq(/rspec|cucumber/i)
       end
 
       it 'sets the logger except filter from a symbol' do
-        described_class.evaluate_guardfile(guardfile_contents: 'logger :except => :jasmine')
+        described_class.evaluate_guardfile(guardfile_contents: 'logger except: :jasmine')
         Guard::UI.options.except.should eq(/jasmine/i)
       end
 
       it 'sets the logger except filter from a string' do
-        described_class.evaluate_guardfile(guardfile_contents: 'logger :except => \'jasmine\'')
+        described_class.evaluate_guardfile(guardfile_contents: 'logger except: \'jasmine\'')
         Guard::UI.options.except.should eq(/jasmine/i)
       end
 
       it 'sets the logger except filter from an array of symbols and string' do
-        described_class.evaluate_guardfile(guardfile_contents: 'logger :except => [:rspec, \'cucumber\', :jasmine]')
+        described_class.evaluate_guardfile(guardfile_contents: 'logger except: [:rspec, \'cucumber\', :jasmine]')
         Guard::UI.options.except.should eq(/rspec|cucumber|jasmine/i)
       end
     end
@@ -290,11 +290,11 @@ describe Guard::Dsl do
       context 'for the log level' do
         it 'shows a warning' do
           Guard::UI.should_receive(:warning).with 'Invalid log level `baz` ignored. Please use either :debug, :info, :warn or :error.'
-          described_class.evaluate_guardfile(guardfile_contents: 'logger :level => :baz')
+          described_class.evaluate_guardfile(guardfile_contents: 'logger level: :baz')
         end
 
         it 'does not set the invalid value' do
-          described_class.evaluate_guardfile(guardfile_contents: 'logger :level => :baz')
+          described_class.evaluate_guardfile(guardfile_contents: 'logger level: :baz')
           Guard::UI.options.level.should eq :info
         end
       end
@@ -302,11 +302,11 @@ describe Guard::Dsl do
       context 'when having both the :only and :except options' do
         it 'shows a warning' do
           Guard::UI.should_receive(:warning).with 'You cannot specify the logger options :only and :except at the same time.'
-          described_class.evaluate_guardfile(guardfile_contents: 'logger :only => :jasmine, :except => :rspec')
+          described_class.evaluate_guardfile(guardfile_contents: 'logger only: :jasmine, except: :rspec')
         end
 
         it 'removes the options' do
-          described_class.evaluate_guardfile(guardfile_contents: 'logger :only => :jasmine, :except => :rspec')
+          described_class.evaluate_guardfile(guardfile_contents: 'logger only: :jasmine, except: :rspec')
           Guard::UI.options.only.should be_nil
           Guard::UI.options.except.should be_nil
         end
@@ -324,24 +324,24 @@ describe Guard::Dsl do
     end
 
     it 'does use the DSL scope plugin' do
-      described_class.evaluate_guardfile(guardfile_contents: 'scope :plugin => :baz')
+      described_class.evaluate_guardfile(guardfile_contents: 'scope plugin: :baz')
       ::Guard.scope[:plugins].should eq [::Guard.plugins(:baz)]
       ::Guard.setup_scope(plugins: [], groups: [])
       ::Guard.scope[:plugins].should eq [::Guard.plugins(:baz)]
     end
 
     it 'does use the DSL scope plugins' do
-      described_class.evaluate_guardfile(guardfile_contents: 'scope :plugins => [:foo, :bar]')
+      described_class.evaluate_guardfile(guardfile_contents: 'scope plugins: [:foo, :bar]')
       ::Guard.scope[:plugins].should eq [::Guard.plugins(:foo), ::Guard.plugins(:bar)]
     end
 
     it 'does use the DSL scope group' do
-      described_class.evaluate_guardfile(guardfile_contents: 'scope :group => :baz')
+      described_class.evaluate_guardfile(guardfile_contents: 'scope group: :baz')
       ::Guard.scope[:groups].should eq [::Guard.groups(:baz)]
     end
 
     it 'does use the DSL scope groups' do
-      described_class.evaluate_guardfile(guardfile_contents: 'scope :groups => [:foo, :bar]')
+      described_class.evaluate_guardfile(guardfile_contents: 'scope groups: [:foo, :bar]')
       ::Guard.scope[:groups].should eq [::Guard.groups(:foo), ::Guard.groups(:bar)]
     end
   end
@@ -358,7 +358,7 @@ describe Guard::Dsl do
       guard :test
     end
 
-    group :x, :halt_on_fail => true do
+    group :x, halt_on_fail: true do
       guard :rspec
       guard :ronn
     end